move ugly code to a separate function
